tap_utils added
