+++ Library/Homebrew/cmd/prune.rb
@@ -1,5 +1,5 @@
 require "keg"
-require "cmd/tap"
+require "tap_utils"
 
 module Homebrew
   def prune
